requirements.txt
flask
pandas
scikit-learn
joblib
requests
streamlit
google-cloud-bigquery
google-cloud-aiplatform
apache-beam[gcp]
